+++ som-interpreter-ast/src/specialized/to_do_node.rs
@@ -23,7 +23,7 @@ impl Invoke for ToDoNode {
 }
 
 impl ToDoNode {
-    fn do_int_loop(start_int: i64, end_int: i64, mut body_block: GCRef<Block>, universe: &mut UniverseAST) -> Return {
+    fn do_int_loop(start_int: i32, end_int: i32, mut body_block: GCRef<Block>, universe: &mut UniverseAST) -> Return {
         let nbr_locals = body_block.borrow().block.borrow().nbr_locals;
         
         for i in start_int..=end_int {
